Apply changes requested by code reviewer
